moved chain specific definitions
