+++ runtime/amplitude/src/definitions.rs
@@ -1,0 +1,4 @@
+
+pub mod asset_hub {
+    pub use runtime_common::parachains::asset_hub::*;
+}
